genai/hf_agents_course/requirements.txt
anthropic==0.52.2
duckduckgo_search==8.0.2
huggingface-hub==0.31.1
ipython==8.18.1
litellm==1.72.2
opentelemetry-exporter-otlp==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-sdk==1.34.0
smolagents==1.17.0
transformers==4.51.3